Teste de Model, Repository e Service do Desejo (Teste de controller ainda não passa)

+++ src/main/java/br/com/insted/funcash/dto/DesejoRequestDTO.java
@@ -11,7 +11,7 @@ import lombok.NoArgsConstructor;
 public class DesejoRequestDTO {
     private String nome;
     private String descricao;
-    private int valor;
+    private double valor;
